+++ src/LSM6DSOX.h
@@ -42,7 +42,7 @@ class LSM6DSOXClass {
     int gyroscopeAvailable(); // Check for available data from gyroscope
 
     // Temperature
-    int readTemperature(int & temperature_deg);
+    int readTemperature(float& temperature_deg);
     int temperatureAvailable();
 
   private:
